Fixes Errors
