+++ room_measurment_tool/room_measurment_tool/ReadMe.txt
@@ -1,0 +1,40 @@
+========================================================================
+    CONSOLE APPLICATION : room_measurment_tool Project Overview
+========================================================================
+
+AppWizard has created this room_measurment_tool application for you.
+
+This file contains a summary of what you will find in each of the files that
+make up your room_measurment_tool application.
+
+
+room_measurment_tool.vcxproj
+    This is the main project file for VC++ projects generated using an Application Wizard.
+    It contains information about the version of Visual C++ that generated the file, and
+    information about the platforms, configurations, and project features selected with the
+    Application Wizard.
+
+room_measurment_tool.vcxproj.filters
+    This is the filters file for VC++ projects generated using an Application Wizard. 
+    It contains information about the association between the files in your project 
+    and the filters. This association is used in the IDE to show grouping of files with
+    similar extensions under a specific node (for e.g. ".cpp" files are associated with the
+    "Source Files" filter).
+
+room_measurment_tool.cpp
+    This is the main application source file.
+
+/////////////////////////////////////////////////////////////////////////////
+Other standard files:
+
+StdAfx.h, StdAfx.cpp
+    These files are used to build a precompiled header (PCH) file
+    named room_measurment_tool.pch and a precompiled types file named StdAfx.obj.
+
+/////////////////////////////////////////////////////////////////////////////
+Other notes:
+
+AppWizard uses "TODO:" comments to indicate parts of the source code you
+should add to or customize.
+
+/////////////////////////////////////////////////////////////////////////////
